requirements.txt
Django==5.2.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
torch
torchvision
Pillow
transformers
langchain
langchain-community
faiss-cpu
sentence-transformers